make tsar use system layout instead of hard coded /usr/local prefix
we introduced a 'make rpm' for easy rpm generating, and may be more
work should be take to get it adapted to other platforms, such as
*BSD etc.

use 'make rpm' to make a rpm from the git tree. and may 'release=10 make
 rpm' to bump the rpm release version as you wish.

anyway, this a huge change in the codes orgnization but not in function
and feature. to outline the change, we bump the version to v2.2.0.